+++ mantle-rpc/Cargo.toml
@@ -18,5 +18,5 @@ url = { version = "1.7", optional = true }
 walrus = { version = "0.8", optional = true }
 
 [features]
-importer = ["saveload", "url", "walrus"]
+import = ["saveload", "url", "walrus"]
 saveload = ["libflate", "serde_json", "failure"]
