+++ mantle-rpc/src/import/mod.rs
@@ -1,6 +1,9 @@
-mod file;
+mod file_importer;
+mod error;
 
 use crate::Interface;
+
+use error::ImporterError;
 
 pub struct Importer {
     backend: Box<dyn ImporterBackend>,
@@ -62,22 +65,4 @@ trait ImporterBackend {
     fn import(&self, name: &str) -> Result<Interface, ImporterError>;
 
     fn import_all(&self) -> Result<Vec<Interface>, ImporterError>;
-}
-
-#[derive(Debug, failure::Fail)]
-pub enum ImporterError {
-    #[fail(display = "Invalid URL: {}", _0)]
-    InvalidUrl(#[fail(cause)] url::ParseError),
-
-    #[fail(display = "No importer for scheme `{}`", _0)]
-    NoImporter(String),
-
-    #[fail(display = "{}", _0)]
-    Fail(#[fail(cause)] failure::Error),
-
-    #[fail(display = "Wasm module missing mantle-interface section")]
-    MissingInterfaceSection,
-
-    #[fail(display = "Could not locate `{}`", _0)]
-    NoImport(String),
 }
